Further implementation of PlayerChildFSM (ChargeMain, Defend)
Nothing concrete yet just some helper functions.

- In DefendState, changed m_fAttackRange to static s_fDetectionRange
- Changed PlayerChildFSM.cs m_currentEnemyCellTarget from type of
Transform to type of EnemyChildFSM so have direct access to script AND
transform.
- Idle State Assigns the script now instead of the transform in the
DetectedEnemyInRange helper function.

Changes In ChargeMain state:
- Execute now checks for presence of targets using the newly
implemented AreThereTargets helper function. Which checks the ECList
from EnemyMainFSM for count > 0.
- Execute then finds targets if there are targets or if the current
target is no longer alive. Alive status is checked using the
IsTargetAlive helper function which checks the the
m_currentEnemyCellTarget CurrentStateEnum == ECState.Dead
- FindTarget helper function Checks an overlap circle for nearby
targets, else it looks through the whole list of enemy child cells and
calculates the closest one and sets m_currentEnemyCellTarget to the
closest one.
- Implementation for when target is alive is NOT DONE YET.

Implemented a Check for deferred state transition method.
- A bool called m_bHasAwaitingDeferredStateChange is used as an
indicator when there is a player command to change states.
- A private PCState in PlayerChildFSM.cs called m_deferredState
contains the state enum which the deferred transition should transition
to.
- Implemented a public void called DeferredChangeState for the player
control script to call. Takes in a PCState enum as parameter. This
method sets the m_bHasAwaitingDeferredStateChange bool indicator and
also the state to transition to.
- Each individual state then checks to see if the bool flag is set to
true at the end of every execute call. If it is, calls the
PlayerChildFSM class’s ExecuteDeferredStateChange function. This will
ensure that the state will not have to stop functions prematurely and
have its state transitioned in the middle of calculations in execute.

Fixed s_nActiveChildCount bug