Completion of EnemyChildCell Defend state and Implementation of Landmine State related scripts
- Completed the defend state of the enemy child cells whereby all the
toggled enemy child cells will be forming a straight defensive line in
front of the enemy child cell

- Implement a PointDatabase script that serve as a library to store
various tactical position that the enemy child cell can browse through
to obtain the most advantegous position to place the landmine. The
position stored in the library will update accordingly every time it is
called (Some of its function still need further testing)

- Implemented a PositionQuery script whereby the Enemychildcell will
send a query toward this script to ask for the best current position to
deploy the landbase. Based on the current game environment and enemy
child cell condition, a decision will be made and send back to the enemy
child cell.

- A new enum called PositionType is added to accommodate to the need of
different type of position, ie. Aggressive, Defensive, Netural

- Implemented a script for debugging usage called ECManualStateControl
whereby the state of the Enemy child can be controlled through keyboard
control of Alpha 1 to Alpha 4

- Enemy child cell folder is further organized