Abandon Reynold's Behavior Approach for Enemy Child Cells major behaviors
- After attempts made, the Reynold's behavior approach (Seperation,
Cohesion, Alignment) will not be used for significant Enemy Child Cells'
major behavior like forming formations. It will be for smaller tasks
such as gathers cells together and spreading them they travel.

- Therefore, to form formation, A FormationDatabase is created. It will
store the alogrithms to calculate the position of various enemy child
cells that must be taken in order to form the formations.

- FormationDatabase will take the in current enemy child cells that are
in the defend state and store them with a specific FormationIndex. This
formation index is then stored with a specific position in the
formation. All of these data are stored in Dictionary data structure
- Enemy Child Cell Manual Cell is changed again to interact with
defending enemy child cell for conviency during debugging process
-  A new SteeringBehavior script in creating to store all the steering
related functions for enemy child cells
-  A new function called GetDefensiveFormation is created in
positionQuery in order for the child cell to determine what defensive
formation to take when defending the main cell.
- A small improvement to the smoothing of A* generated path whereby it
will remove uneccessory waypoints in the generated path
- Two rotating functions are added, RotateToHeading and RotateRandom.
One will rotate based its velocity vector and face the direction it is
traveling to, another will be simply rotating clockwise and
anticlockwise randomly.
- All the attack-related enemy child cells states are implemented with
steering behavior and rotating functions
- Added a new Enum called "Formation" into the Enum script