Official Implementation of 'Defence' State for Squad Groups
*THIS COMMISSION EDITED THE SCENE FILE!*

> SquadChildFSM.cs, SC_States.cs, PSLogicaliser.cs
- Implementation of defence system. Simply spawn the Squad Captain to
see the effect.
- The effectivness of the defence mechanism is dependent on the number
of cells. (If there are lesser cells, they will be more spaced out
amongst each other thus more penetratable for the enemies)
- Currently, 75% of the cells will automatically assign to defence state
when there is more than 10 cells in production states.

NOTE: If you think that the defence system is too op for the game,
simply:
Go to squad captain gameobject in the scene, locate PlayerSquadFSM.cs in
the inspector and INCREASE fCooldownMinimum and fCooldownMaximum. This
increases the spawn delay for the cells.