MAJOR CHANGE: Fixed Node List bugs
- Changed Node_Manager List system to that of array.
- activeChildCount from Node_manager still works.
- PlayerChildFSM now has a getter to access the array of playerChildFSM
object pool
- PlayerChildFSM also has another array of enum of pcStatus. This array
will correspond to the status of the PlayerChildFSM of the same index
- Node_manager still handles adding and removing child cells from the
lists. It will manage the pcStatus array in PlayerChildFSM.
- Node_manager now has a SendChildToAttack function as well which
instead of setting the pcStatus array to InLeftNode/InRightNode or
DeadState, it will set it to attacking so we can know that it is
attacking and we shouldn’t touch it until it returns to DeadState. We
also know that it is no longer in the nodes and hence we can’t call it
back.

Misc:
- Turned the GetNodeChildList to DEPRECIATED_GetNodeChildList to avoid
terrible crashes. It currently returns a very inefficiently produced
List of the child cells in that node
- Turned all other instances of
GetNode(_selectedNode).GetNodeChildList().Count to
GetNode(_selectedNode).activeChildCount