Plan-B Attacking Strategy on Enemy Landmines
When producing, squad children will now avoid or attack the incoming
landmines based on the number of landmines. If there are more landmines
than producers, the squad children will avoid the landmines, else
certain amount of squad children will sacrifice themselves the greater
good of squad kind.

> Enum.cs
- Added 'Avoid' state to both Squad Captain FSM and Squad Child FSM

> PlayerSquadFSM.cs
- Fixed fertility problem -> There is squad children in production but
the squad captain is not producing

> PS_Logicaliser.cs
- Removed whole logic, Squad Captain is pretty dumb right now

> PS_States.cs & SC_States.cs
- Added 'Avoid' state
- Added logic for 'Avoid' state

> SquadChildFSM.cs
- Moved static variables to its desired place
- Added a method and a static list to handle enemy's landmine list