Implemented getters for player child cells in specific states
- Now there are quick-access getters for player child cells that are
re-calculated every fixed update.
- childrenInLeftNode, childrenInRightNode, childrenInAttack.
- The array returns a sized maxPlayerChildCount + 1 int array
containing the poolIDs of each individual child cell. This would allow
you to iterate through the array in order to quickly grab each
PlayerChildFSM that fulfils your criteria.
- To know when the array has ended, simply check for the marker -1. -1
value would indicate that you have iterated to the end of the returned
array and that there are no more child cells in this status.